refactor(client): remove unnecessary runtime
