+++ jina/peapods/peas/__init__.py
@@ -132,7 +132,7 @@ class BasePea(metaclass=PeaMeta):
 
         # TODO: BAD pattern here!
         #   It is unnecessary to do "`a` in b and b.a" for `Namespace`, the only reason
-        #   for this check is because PyClient is using (abusing?) `BasePea` but PyCilent's args is not
+        #   for this check is because Client is using (abusing?) `BasePea` but PyCilent's args is not
         #   generated by set_pea/pod_parser(). Therefore many information are missing and make this check
         #   necessary.
         #   (Han 2020.12.12)
