+++ tests/unit/test_helper.py
@@ -2,12 +2,11 @@ import random
 import time
 from types import SimpleNamespace
 
-import numpy as np
 import pytest
 
 from cli import _is_latest_version
 from jina import NdArray, Request
-from jina.clients.python.helper import safe_callback, pprint_routes
+from jina.clients.helper import safe_callback, pprint_routes
 from jina.drivers.querylang.queryset.dunderkey import dunder_get
 from jina.excepts import BadClientCallback
 from jina.helper import cached_property, convert_tuple_to_list, complete_path
